yeah, here comes TinyQ :) read the README for notes about it.
libyzis and nyzis have been ported to tinyq.
cyzis is disabled for now.
I removed all C bindings since it appears we wont work on that in the near future.

YZLine is basically a QString now.
It compiles and runs for me except nyzis does not properly open the test file.
Prolly some memory problem in allocating YZLines. Maybe we can change that to a more QtLike stuff.

And :

						*********** Happy Birthday Thomas :) ****************

Mik