ok , here is the second try :) everything is libtoolised now.
we use a fake libtool lib in libyzis to build 2 others libs : libqtyzis and libtinyqyzis
first one is linked with qt the other with tinyq.

nyzis uses libtinyqyzis as a shared lib (can be static easily i think)
and kyzis uses libqtyzis as shared.

kyzis and nyzis both work here.