Big commit to add some tests to Yzis.
I provide the test library inside the tests, to help people run the tests
easily. You can find the original at cppunit.sf.net

This is still pretty rough, for example it uses qmake to compile the test
library and the tests itself. One advantage is that it compiles nicely on both
KDE and Windows but it does not integrate into the rest of Yzis compilation
scheme.

There are still many tests to write and run.