adding my thoughts about motions support
what is a motion ?
:help motion in vim :)

or
try to reproduce the command : d3w in yzis
this should delete the next 3 words
in Vim, there is also : sentences, paragraphs, ....

we could just hardcode it for every functions but this is ugly :)
so I am trying to define a motion support framework
so that we could say :
'd' okay it's a delete command, what should I do exactly ? then I look at the motion, and let someone else calculate all
the stuff and give me back "what I should remove"

my current toughts is to make a mostly static class (YZMotion) which will be initially fed
with a set a RegExps to detect motions in commands.
When we detected one motion, we let the "command function" (the function which gets called from commands.cpp) call an
YZMotion static function to calculate whatever we need (could be a length or a position in text). So we can
centralize all the dirty code to calculate lengths and positions using regexps etc ...

any comments welcome