This is a big patch but the functional change is small.
The content of a line of buffer is now accessible with a method textline()
instead of data(). Using textline() makes the code more readable and provides
an easy way to distinguish between the line content (text) and the future line
meta-data (syntax highlighting, ...)

A method setTextLine() is also added but is protected. The method provides
correct encapsulation and is easier to read. Previously, the line was updated
with at(line)->setData() which is not optimal in term of maintainability. We
can now transparentely switch the structure of yzline, or even suppress the
metadata.

setTextLine() is a good place to recalculate later the syntax highlighting
when the text changes.

A method yzline() is replaces the previous at() for the same reason as above.
The method is protected, as it should be.

If you have uncommitted code, you have to replace the data() with textline()
to make it compile again.