reimporting tests
I have dropped the cppunit library we were distributing before, so you need to install cppunit >= 1.8.0
to be able to run the tests
it is detected by configure, if it's not found , the tests/ directory cannot be compiled

we have a libphilcppunit library to use Phil's functions, it is built but not installed
in the tests/ directory the binary tests is built and can be run.
you can use 'env_tests' to make use you use the 'just built' libqtyzis.