- add drawPool system : instead of sending several paintEvent at the same time to the gui, use : setPaintAutoCommit( false ); sendPaintEvent( ... ); .. commitPaintEvent();
setPaintAutoCommit( bool ) -> auto commit to the gui when calling sendPaintEvent
sendPaintEvent( .. ) -> if auto commit, send a paintEvent to the gui, else cumulate areas to be repainted.
removePaintEvent( ... ) -> remove an area from the cumulated ones.
commitPaintEvent(); -> send paintEvents to the gui according to the cumulated events.

- visual mode now uses drawPool

CCMAIL: 51-done@bugs.yzis.org