Here comes Qt4 support for libyzis and nyzis you need : Qt4 > beta1 (strictly superior since it has a bug ;), today's snapshot works)
there are a lot of warnings with qt4, this is normal we will remove them when we switch completely to Qt4
(removing them now creates new warnings in Qt3 :)

when you add new code, try to make sure it is qt4 compatible (it's pretty hard ...)
or #ifdef-it properly please :)

there are still some bugs lying around but I'll fix them
it should not change anything in the Qt3 build, if it does please complain to me ;)

btw, I don't want to see -lQt3Compat in libqt4yzis, thanks :) (easy way to not do things that need to be done)
it's even _useless_ since you have to rename classes anyway.

in Qt3 we used to rely on .setAutoDelete(true) for some parts. This disappeared in Qt4. We will need to review this code to make sure we properly clean up memory when needed.

Enjoy,
Mik