Qt4 porting is already underway, but this change should simplify things.  Instead of using QValueList or QValueVector, we can now use YZList and YZVector.  These classes inherit from the correct Qt container based on the Qt version number.
There are two ways of achieving this, one is with #define and the other is subclassing.  I chose subclassing because of its type-safety and it's harder to create subtle compile errors that the preprocessor is prone to do.

That said, I'm not dead set on the approach, if others think #define is the way to do it, by all means change it.  The important thing is keeping /trunk and /qt4_branch as close as possible.

 r295@laptop:  craig | 2005-08-03 16:20:44 -0400
 Create YZVector and YZList to solve Qt4 compatibility problems.